moving more error handling, still haven't gotten middleware working
