+++ model-engine/model_engine_server/common/datadog_utils.py
@@ -10,6 +10,15 @@ def add_trace_resource_name(tag: str):
         current_span.set_tag("launch.resource_name", tag)
 
 
+def add_trace_request_id(request_id: str):
+    """Adds a custom tag to a given dd trace corresponding to the request id
+    so that we can filter in Datadog easier
+    """
+    current_span = tracer.current_span()
+    if current_span:
+        current_span.set_tag("launch.request_id", request_id)
+
+
 def get_request_id():
     """Gets the request id for an api call (in our case, dd trace id) so that we can filter in Datadog easier"""
     current_span = tracer.current_span()
