+++ model-engine/model_engine_server/domain/use_cases/llm_fine_tuning_use_cases.py
@@ -13,7 +13,7 @@ from model_engine_server.common.dtos.llms import (
     ListFineTunesResponse,
 )
 from model_engine_server.core.auth.authentication_repository import User
-from model_engine_server.core.domain_exceptions import ObjectNotFoundException
+from model_engine_server.domain.exceptions import ObjectNotFoundException
 from model_engine_server.core.loggers import filename_wo_ext, make_logger
 from model_engine_server.domain.entities import BatchJobStatus
 from model_engine_server.domain.exceptions import InvalidRequestException, LLMFineTuningQuotaReached
