+++ model-engine/model_engine_server/infra/repositories/db_repository_mixin.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from functools import wraps
 from typing import Callable
 
-from model_engine_server.core.domain_exceptions import ReadOnlyDatabaseException
+from model_engine_server.domain.exceptions import ReadOnlyDatabaseException
 from sqlalchemy.ext.asyncio import AsyncSession
