+++ model-engine/tests/unit/infra/repositories/test_db_model_endpoint_record_repository.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, Mock
 
 import pytest
 from model_engine_server.common.dtos.model_endpoints import ModelEndpointOrderBy
-from model_engine_server.core.domain_exceptions import ReadOnlyDatabaseException
+from model_engine_server.domain.exceptions import ReadOnlyDatabaseException
 from model_engine_server.db.models import Bundle, Endpoint
 from model_engine_server.domain.entities import ModelEndpointRecord
 from model_engine_server.infra.gateways import FakeMonitoringMetricsGateway
