+++ model-engine/tests/unit/infra/services/test_live_batch_job_orchestration_service.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 import pytest
 from model_engine_server.common.constants import DEFAULT_CELERY_TASK_NAME, LIRA_CELERY_TASK_NAME
 from model_engine_server.common.dtos.tasks import GetAsyncTaskV1Response, ResponseSchema, TaskStatus
-from model_engine_server.core.domain_exceptions import ObjectNotFoundException
+from model_engine_server.domain.exceptions import ObjectNotFoundException
 from model_engine_server.domain.entities import (
     BatchJob,
     BatchJobSerializationFormat,
