+++ model-engine/tests/unit/infra/services/test_live_endpoint_builder_service.py
@@ -8,7 +8,7 @@ from model_engine_server.common.dtos.endpoint_builder import (
     BuildEndpointResponse,
     BuildEndpointStatus,
 )
-from model_engine_server.core.domain_exceptions import DockerBuildFailedException
+from model_engine_server.domain.exceptions import DockerBuildFailedException
 from model_engine_server.core.fake_notification_gateway import FakeNotificationGateway
 from model_engine_server.core.notification_gateway import NotificationApp
 from model_engine_server.domain.entities.model_bundle_entity import (
